// app/verticals/commerce-customer-context/src/api/retail-purchasing-subject-current-v1.read.ts
export interface RetailPurchasingSubjectCurrentV1Services {
  readonly retailPrincipalResolution: RetailPrincipalResolutionServices;
}